intl/l10n/test/dom/test_domloc_attr_sanitized.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test DOMLocalization's attr sanitization functionality</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript">
  "use strict";
  const { DOMLocalization } =
    ChromeUtils.import("resource://gre/modules/DOMLocalization.jsm", {});
  const { MessageContext } =
    ChromeUtils.import("resource://gre/modules/MessageContext.jsm", {});

  async function* mockGenerateMessages(locales, resourceIds) {
    const mc = new MessageContext(locales);
    mc.addMessages(`
key1 = Value for Key 1

key2 = Value for <a>Key 2<a/>.
    `);
    yield mc;
  }

  SimpleTest.waitForExplicitFinish();
  addLoadEvent(async () => {
    const domLoc = new DOMLocalization(
      [],
      mockGenerateMessages
    );

    await domLoc.translateFragment(document.body);

    const elem1 = document.querySelector("#elem1");
    const elem2 = document.querySelector("#elem2");

    ok(elem1.textContent.includes("Value for"));
    // This is a limitation of us using Node.localize API
    // Documenting it here to make sure we notice when we fix it
    is(elem1.getAttribute("title"), "Old Translation");

    ok(elem2.textContent.includes("Value for"));
    ok(!elem2.hasAttribute("title"));

    SimpleTest.finish();
  });
  </script>
</head>
<body>
  <p id="elem1" title="Old Translation" data-l10n-id="key1"></p>
  <p id="elem2" title="Old Translation" data-l10n-id="key2"></p>
</body>
</html>
